# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: OpenDesk Remote API
  version: 0.1.0
  description: REST API skeleton for OpenDesk Remote control plane.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Signed session bearer token backed by a server-side session hash, or user-bound OpenDesk API token.
    sessionCookie:
      type: apiKey
      in: cookie
      name: opendesk_session
paths:
...
  /api/v1/address-books:
    get:
      summary: List address books
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200": { description: Address books returned }
        "401": { description: Authentication required }
    post:
      summary: Create address book
      security:
        - bearerAuth: []
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                description: { type: string }
                owner_user_id: { type: integer }
                entries:
                  type: array
                  items:
                    type: object
                    required: [device_id]
                    properties:
                      device_id: { type: integer }
                      alias: { type: string }
      responses:
        "201": { description: Address book created }
        "400": { description: Invalid address book }
        "401": { description: Authentication required }
...
  /api/v1/api-tokens:
    get:
      summary: List API tokens
      security:
        - bearerAuth: []
        - sessionCookie: []
      responses:
        "200": { description: API tokens returned without raw token or token hash }
        "401": { description: Authentication required }
    post:
      summary: Create API token
      description: Returns the raw API token once. Persistence stores only a hash.
      security:
        - bearerAuth: []
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, scopes]
              properties:
                name:
                  type: string
                scopes:
                  type: array
                  items: { type: string }
                expires_at:
                  type: string
                  format: date-time
      responses:
        "201": { description: API token created and raw token returned once }
        "400": { description: Invalid API token request }
        "401": { description: Authentication required }
...